py-smelt/pyproject.toml
[project]
name = "pysmelt"
version = "0.3.0"
description = "smelt is a system to describe tests for eda"
requires-python = ">=3.8"
dependencies = [
  "typer>=0.9.0",
  "pyyaml>=6.0.1",
  "pydantic>=2.6.0",
  "toml>=0.10.2",
  "betterproto[compiler]>=1.2.5",
  "rich>=13.7.1",
]
[[project.authors]]
name = "James Connolly"
email = "james@silogy.io"

[build-system]
requires = ["maturin>=1.5,<2.0"]
build-backend = "maturin"

[project.scripts]
smelt = "pysmelt.cli:main"

[tool.maturin]
features = ["pyo3/extension-module"]
